compare_sizes: refactor
