+++ .gitignore
@@ -17,3 +17,6 @@ neardev
 .idea
 .vscode
 **/.DS_Store
+
+# Python
+__pycache__
